+++ agent/src/main/java/com/cmclinnovations/agent/VisBackendAgent.java
@@ -180,6 +180,7 @@ public class VisBackendAgent {
   @GetMapping("/form/{type}")
   public ResponseEntity<Map<String, Object>> getFormTemplate(@PathVariable(name = "type") String type) {
     LOGGER.info("Received request to get the form template for {}...", type);
+    // Access to this empty form is prefiltered on the UI and need not be enforced
     return this.getService.getForm(type, false, new HashMap<>());
   }