+++ agent/src/main/java/com/cmclinnovations/agent/service/GetService.java
@@ -33,7 +33,8 @@ public class GetService {
    * @param kgService            KG service for performing the query.
    * @param queryTemplateService Service for generating query templates.
    */
-  public GetService(KGService kgService, QueryTemplateService queryTemplateService) {
+  public GetService(KGService kgService,
+      QueryTemplateService queryTemplateService) {
     this.kgService = kgService;
     this.queryTemplateService = queryTemplateService;
   }
@@ -66,8 +67,7 @@ public class GetService {
     }
     String query = this.queryTemplateService.getShaclQuery(resourceID, requireLabel);
     Queue<Queue<SparqlBinding>> nestedVariablesAndPropertyPaths = this.kgService.queryNestedPredicates(query);
-    return this.getInstances(nestedVariablesAndPropertyPaths, targetId, parentField,
-        addQueryStatements, addVars);
+    return this.getInstances(nestedVariablesAndPropertyPaths, targetId, parentField, addQueryStatements, addVars);
   }
 
   /**
@@ -125,7 +125,9 @@ public class GetService {
       // Removes the first instance from results as the core instance
       SparqlBinding firstInstance = results.poll();
       // Iterate over each result binding to append arrays if required
-      results.stream().forEach(firstInstance::addFieldArray);
+      results.stream().forEach(result -> {
+        firstInstance.addFieldArray(result, new HashMap<>());
+      });
       return firstInstance;
     }
     if (results.size() == 1) {
@@ -196,7 +198,8 @@ public class GetService {
     Queue<SparqlBinding> instances = this.kgService.query(getQuery.poll(), SparqlEndpointType.MIXED);
     // Query for secondary instances ie instances that are subclasses of parent
     Queue<SparqlBinding> secondaryInstances = this.kgService.query(getQuery.poll(), SparqlEndpointType.BLAZEGRAPH);
-    instances = this.kgService.combineBindingQueue(instances, secondaryInstances);
+    instances = this.kgService.combineBindingQueue(instances, secondaryInstances,
+        this.queryTemplateService.getArrayVariables());
     // If there is a variable sequence available, add the sequence to each binding,
     if (!varSequence.isEmpty()) {
       instances.forEach(instance -> instance.addSequence(varSequence));